Updated to new request and service system
Requests are now objects that can be saved and executed instead of needing to call functions on the service directly. Each request holds information on the client service that it belongs to, request parameter information, and functions for executing the request synchronously and asynchronously. Client Services now contain generic information needed for each request and also can be used for handling HttpResponse messages and error handling if needed.