Created Message Processing and Subscription System
A Message Sink is responsible for receiving incoming messages. A Subscription Service is responsible for allowing consumers to subscribe and unsubscribe to different event types and when that event type is triggered, executing the callback function provided. The Message Processor class handles both of these things by allowing producers send events to it and consumers subscribe and listen for those events.