Removed Ludwig Commands / Modified Notify Youtube Playlist Upload Commands
Removed methods pertaining to just Ludwig as these things can be done through the more generalized commands. Modified the Notify commands to now just check against the database and see if we can add or remove a notification task instead of just checking the task handler. After any database changes, a Timed Task Message is then sent to our message sink. The Timed Task Handler is subscribed to listen of messages of type TimedTaskMessage and add or remove associated tasks.