Removed Service Request Initializer / Made Requests and Resource Internal
Removed the service request initializer and instead now pass in just the IClientService to the request. Each request now just uses an abstract property for determining what type of HttpMethod to use for the request. Each Request and Resource has now been made internal and sealed so that ClientServices have direct control over them and they cannot be instantiated outside of the ClientService.