Added Try Catch for Timed Task Exceptions
For now, the Timed Task Handler will wrap task handling in a try catch block to catch any errors that might occur from a singular task. This probably needs to be converted to work a different way in the future but for now this should keep the timed task handler from crashing everything if one task fails.