Refactored Task Handler
Refactored task handler to now take in a list of ITimedTasks to be initialized with. Changed the task handler to be a hosted service. The task handler now also uses a concurrent bag for adding new tasks instead of locking and awaiting a lock.

Added ITimedTaskRetriever interface for getting a timed task enumerable. This is so we can retrieve a set of timed tasks without needing to know the implementation details of how those tasks came to be. As an example, YoutubeTimedTaskDatabaseRetriever returns a set of new TimedTask instances with information gathered from the database.